% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{AddParPed}
\alias{AddParPed}
\title{Partial pedigree fix}
\usage{
AddParPed(PedIN)
}
\arguments{
\item{PedIN}{A dataframe with IDs in the first column and parents in the
second and third columns}
}
\value{
A dataframe with one row per individual occuring in the pedigree
}
\description{
Add rows for all parents in pedigree who are not yet represented in the
first column
}

