% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMaybeRel.R
\name{CalcSnpdBoth}
\alias{CalcSnpdBoth}
\title{SNPs scored in both individuals}
\usage{
CalcSnpdBoth(Pairs, GenoM)
}
\arguments{
\item{Pairs}{dataframe or matrix with 2 columns with IDs}

\item{GenoM}{matrix with genotype data, individuals in rows and SNPs in columns,
missingness as -9}
}
\value{
an unnamed vector with counts
}
\description{
count no. SNPs for which both individuals of pair are
  successfully SNPd
}
\keyword{internal}
