% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{CheckAP}
\alias{CheckAP}
\title{check AgePrior}
\usage{
CheckAP(AgePrior)
}
\arguments{
\item{AgePrior}{matrix with 'MaxAgeParent' rows and columns M-P-FS-MHS-PHS}
}
\value{
AgePrior in corrected format, if necessary
}
\description{
Check that the provided AgePrior is in the correct format
}
\keyword{internal}
