% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{CheckGeno}
\alias{CheckGeno}
\title{check GenoM}
\usage{
CheckGeno(GenoM)
}
\arguments{
\item{GenoM}{the genotype matrix}
}
\value{
updates GenoM internally to remove low call rate individuals (<.5)
  and low call rate SNPs (<0.1), if necessary
}
\description{
Check that the provided genotype matrix is in the correct format
}
\keyword{internal}
