% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkAgePrior.R
\name{CountAgeDif}
\alias{CountAgeDif}
\title{Tabulate Age Differences}
\usage{
CountAgeDif(BirthYear, BYrange = range(BirthYear))
}
\arguments{
\item{BirthYear}{numeric vector with birth years.}

\item{BYrange}{range to limit counts to.}
}
\description{
Count no. pairs per age difference from birth years. Quicker
  than \code{table(outer())}.
}
\keyword{internal}
