% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCompare.R
\name{DyadCompare}
\alias{DyadCompare}
\title{Compare dyads}
\usage{
DyadCompare(Ped1 = NULL, PedFile1 = NULL, Ped2 = NULL, PedFile2 = NULL,
  sep1 = "\\t", sep2 = "")
}
\arguments{
\item{Ped1}{Original pedigree, dataframe with 3 columns: id-dam-sire}

\item{PedFile1}{filename of original pedigree, only the first 3 columns
will be used. Provide EITHER Ped1 OR PedFile1.}

\item{Ped2}{Infered pedigree, e.g. SeqOUT$Pedigree}

\item{PedFile2}{filename of infered pedigree}

\item{sep1}{the field separator character in 'PedFile1'}

\item{sep2}{the field separator character in 'PedFile2'}
}
\value{
A 3x3 table with the number of pairs assigned as full siblings (FS),
  half siblings (HS) or unrelated (U) in the two pedigrees.
}
\description{
Count the number of half and full sibling pairs correctly and incorrectly
assigned
}
\seealso{
\code{\link{PedCompare}}
}

