% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{eqv}
\alias{eqv}
\title{Comparison}
\usage{
eqv(x, V)
}
\arguments{
\item{x}{a value}

\item{V}{a vector of the same type as x}
}
\value{
a logical vector of the same length as v
}
\description{
Test which elements in a vector are equal to x, returning FALSE at missing
values in V
}

