% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{FacToNum}
\alias{FacToNum}
\title{Convert factor to numeric}
\usage{
FacToNum(x)
}
\arguments{
\item{x}{a factor}
}
\value{
A numeric vector of the same length as x.
}
\description{
Converts factors to their numeric values, via \code{as.character}.
}

