% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{FieldMums_griffin}
\alias{FieldMums_griffin}
\title{Example field-observed mothers: griffins}
\format{
A data frame with 144 rows and 2 variables (id, mum)
}
\usage{
data(FieldMums_griffin)
}
\description{
Example field pedigree used in vignette for
  \code{\link{PedCompare}} example. Non-genotyped females have IDs 'BlueRed',
  'YellowPink', etc.
}
\seealso{
\code{\link{SeqOUT_griffin}} for a sequoia run on simulated genotype
  data, \code{\link{Ped_griffin}} for the 'true' pedigree.
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
