% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedToNum.R
\name{FoldSibGPs}
\alias{FoldSibGPs}
\title{Fold IDs of Sibship Grandparents}
\usage{
FoldSibGPs(PedNum, Ng, Nd)
}
\arguments{
\item{PedNum}{pedigree, ids replaced by numbers, dummies negative.}

\item{Ng}{no. genotyped indivs.}

\item{Nd}{length 2 vector, no. female & male dummies.}
}
\value{
An integer vector, with missing values as 0.
}
\description{
Fold IDs of sibship grandparents into a 2 x nInd/2 x 2 array, as
  they are stored in Fortran, and then stretch this into a vector that can be
  passed to Fortran and easily be transformed back into said 3D array.
}
\keyword{internal}
