% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{\%ina\%}
\alias{\%ina\%}
\title{Value Matching}
\usage{
x \%ina\% y
}
\arguments{
\item{x}{vector: the values to be matched}

\item{y}{vector: the values to be matched against}
}
\value{
A logical vector of the same length as x.
}
\description{
Like \code{\%in\%}, returns a logical vector indicating if there is a match
 or not for its left operand, but returns NA for NA's in the left operand.
}

