% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{function in Examples from integer {base}}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a number}

\item{tol}{tolerance}
}
\description{
function in Examples from integer {base}
}

