% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcParentProbs.R
\name{LLtoProb}
\alias{LLtoProb}
\title{transform log-likelihoods to probabilities}
\usage{
LLtoProb(LLv)
}
\arguments{
\item{LLv}{a vector with log10-likelihoods. All values >0 are set to NA.}
}
\value{
a vector with probabilities, with the same length and names.
}
\description{
transform a vector with log10 likelihoods to a vector with
 probabilities summing to one.
}
\details{
Computational under/overflow issues are reduced by subtracted the
  maximum value before converting from log to regular scale. Probabilities
  that would still be smaller than the machine precision (\code{(LL -
  min(LL)/2) < log10(.Machine$double.xmin)}) are set to NA en then to 0,
  instead of \code{-Inf}, to avoid issues when scaling to sum to 1.
}
