% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_F90wrappers.R
\name{NumToID}
\alias{NumToID}
\title{replace numbers by names}
\usage{
NumToID(x, k = NULL, realID = NULL, dumID = NULL)
}
\arguments{
\item{x}{vector with numbers}

\item{k}{1=maternal, 2=paternal (for dummies only)}

\item{realID}{vector with IDs, order as in genotype file}

\item{dumID}{2-column matrix with dummy IDs}
}
\value{
character vector with IDs
}
\description{
Replace the rownumbers by IDs, as in the genotype file, and replace negative
numbers by dummy IDs
}

