% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_F90wrappers.R
\name{orderLH}
\alias{orderLH}
\title{Order lifehistory data}
\usage{
orderLH(LH = NULL, gID = NULL)
}
\arguments{
\item{LH}{dataframe with lifehistory information:
\itemize{
 \item{ID: }{max. 30 characters long,}
 \item{Sex: }{1 = females, 2 = males, other numbers = unkown,}
 \item{Birth Year: }{(or hatching year) Use negative numbers to denote
 missing values.}}}

\item{gID}{character vector with IDs in genotype data, in order of occurence}
}
\value{
\item{BY}{Numeric vector with birth years, of same length as gID}
\item{Sex}{Numeric vector with genders, of same length as gID}
}
\description{
Order lifehistory data to match order of IDs in genotype data,
  filling in gaps with missing values
}
\keyword{internal}
