% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{pasteD}
\alias{pasteD}
\title{paste directory and file name}
\usage{
pasteD(folder, fileName)
}
\arguments{
\item{folder}{foldername}

\item{fileName}{filename}
}
\description{
paste directory and file name
}

