% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Ped_griffin}
\alias{Ped_griffin}
\title{Example pedigree: griffins}
\format{
A data frame with 200 rows and 4 variables (id, dam, sire, birthyear)
}
\usage{
data(Ped_griffin)
}
\description{
Example pedigree used in the ageprior vignette, with overlapping
  generations.
}
\seealso{
\code{\link{LH_griffin}}; \code{\link{SeqOUT_griffin}} for a sequoia
  run on simulated genotype data based on this pedigree;
  \code{\link{Ped_HSg5}} for another pedigree, \code{\link{sequoia}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
