% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{Ped_HSg5}
\alias{Ped_HSg5}
\title{Example pedigree}
\format{
A data frame with 1000 rows and 3 variables (id, dam, sire)
}
\usage{
data(Ped_HSg5)
}
\description{
This is \strong{Pedigree II} in the paper, with discrete
  generations and considerable inbreeding
}
\references{
Huisman, J. (2017) Pedigree reconstruction from SNP data:
  Parentage assignment, sibship clustering, and beyond. Molecular Ecology
  Resources 17:1009--1024.
}
\seealso{
\code{\link{LH_HSg5} \link{SimGeno_example} \link{sequoia}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
