% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkAgePrior.R
\name{PlotAgePrior}
\alias{PlotAgePrior}
\title{Plot age priors}
\usage{
PlotAgePrior(AP = NULL, APlist = NULL)
}
\arguments{
\item{AP}{matrix with age priors of dimension nAgeclasses by 9, as ' AgePriors' in Sequoia output}

\item{APlist}{output list from MakeAgePriors when called with Return='all', i.e. including
various intermediate output.}
}
\description{
visualise the age-difference based prior probability ratios
}
\seealso{
\code{\link{MakeAgePrior}}
}
