% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePairs.R
\name{rc}
\alias{rc}
\title{Find siblings}
\usage{
rc(x, Ped)
}
\arguments{
\item{x}{an ID}

\item{Ped}{a pedigree with columns id - dam - sire}
}
\value{
The individuals which are full or half siblings to x, as a
  three-column matrix with column names id1 (x), id2 (the siblings), and
  RC (the relatedness category, 'FS' or 'HS').
}
\description{
Find siblings
}
\keyword{internal}
