% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ReadSpecs}
\alias{ReadSpecs}
\title{Read specs file}
\usage{
ReadSpecs(SpecsFile = "SequoiaSpecs.txt")
}
\arguments{
\item{SpecsFile}{comma-separated text file with parameter settings,
containing 1 column with labels and 1 column with values}
}
\description{
Read the sequoia specs text file.
}

