% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ReadTable}
\alias{ReadTable}
\title{Data input}
\usage{
ReadTable(..., sep = "\\t")
}
\arguments{
\item{...}{parameters for read.table}

\item{sep}{column seperator}
}
\description{
Read data with header row.
}

