% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_F90wrappers.R
\name{SeqParents}
\alias{SeqParents}
\title{Parentage assignment}
\usage{
SeqParents()
}
\value{
A dataframe with the number and proportion of genotyped individuals
  with a female and male parent assigned. The assigned parents are written
  to a text file, by default `ParentsAssigned.txt'.
  In addition, it will create a text file with parent-offspring pairs where
  it is unclear which is the parent and which is the offspring, as either
  individual does not have a birthyear, and inferred sex for individuals
  with no sex provided, but which could be assigned as parent.
}
\description{
Assign genotyped parents to genotyped individuals. Parameter values and
filenames are taken from the SequioiaSpecs text file, generated by
\code{\link{SeqPrep}}.
}
\details{
The output file, by default `ParentsAssigned.txt', contains the following
columns:
\describe{
  \item{id}{Individual ID}
  \item{dam}{Assigned mother, or NA}
  \item{sire}{Assigned father, or NA}
  \item{LLR_dam}{Log-Likelihood Ratio (LLR) of this female being the mother,
     versus the next most likely relationship between the focal individual and
     this female (see ... for relationships considered)}
  \item{LLR_sire}{idem, for male parent}
  \item{LLR_pair}{trio LLR, versus the next most likely configuration
     (containing one or neither parent)}
  \item{OH_dam}{Number of loci at which the offspring and mother are
     opposite homozygotes}
  \item{OH_sire}{idem, for male parent}
  \item{RowO}{Offspring's row identifier in the genotype file. Used by
     \code{link{SeqSibs}}, please do not modify}
  \item{RowD}{Dam's row identifier}
  \item{RowS}{Sire's row identifier}
}
}

