% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_F90wrappers.R
\name{SeqSibs}
\alias{SeqSibs}
\title{Sibship clustering}
\usage{
SeqSibs()
}
\value{
A dataframe with the number and proportion of individuals with a
female and male genotyped or dummy parent assigned. The pedigree is written
to a text file, by default `PedSeq.txt'.
}
\description{
Create clusters of full- and half- siblings with a dummy parent,
assign grandparents to such sibships, and assign parental relationships
between dummies. Parameter values and filenames are taken from the
SequioiaSpecs text file, generated by \code{\link{SeqPrep}}.
}
\details{
The output file, by default `PedSeq.txt', contains the following columns:
\describe{
  \item{id}{Individual ID}
  \item{dam}{Assigned mother, or NA}
  \item{sire}{Assigned father, or NA}
  \item{LLR_dam}{Log-Likelihood Ratio (LLR) of this female being the mother,
     or the individual belonging to this sibship, versus the next most
     likely relationship between the focal individual and this female or
     sibship (see ... for relationships considered)}
  \item{LLR_sire}{idem, for male (dummy) parent}
  \item{LLR_pair}{trio LLR, versus the next most likely configuration
     (containing one or neither (dummy) parent)}
}
}

