% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparePeds.R
\name{SibMatch}
\alias{SibMatch}
\title{Find the closest matching inferred sibship to a true sibship}
\usage{
SibMatch(SimX, Infrd, SNPd)
}
\arguments{
\item{SimX}{a vector with the IDs in the true (Ped1) sibship}

\item{Infrd}{a list of vectors with the IDs in the inferred (Ped2) sibships}

\item{SNPd}{character vector with IDs of genotyped individuals}
}
\value{
a named numeric vector with the number of matches ('NumMatch'),
  the position of the best match ('Best'), the inferred sibship size of
  this best match ('Tot'), the number of matching IDs ('OK'), and the
  number of mismatches ('err').
}
\description{
Find the closest matching inferred sibship to a true sibship
}
\keyword{internal}
