% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{SimGeno_example}
\alias{SimGeno_example}
\title{Example genotype file}
\format{A data frame with 214 rows and 201 columns: id, followed by 1 column
per SNP coded as 0/1/2 or -9 for missing values.}
\usage{
data(SimGeno_example)
}
\description{
Simulated genotype data for cohorts 1+2 in Pedigree Ped_HSg5
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\references{
Huisman, J. Pedigree reconstruction from SNP data: Parentage
  assignment, sibship clustering, and beyond. (under review) Molecular
  Ecology Resources
}
\seealso{
\code{\link{Ped_HSg5} \link{sequoia}}
}
\keyword{datasets,}
\keyword{sequoia}

