% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{supprOF}
\alias{supprOF}
\title{Wrapper for suppressMessages}
\usage{
supprOF(X, quiet = FALSE)
}
\arguments{
\item{X}{a function}

\item{quiet}{suppress messages / not}
}
\description{
Provide additional on/off argument to suppressMessages()
}

