% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Table}
\alias{Table}
\title{table}
\usage{
Table(...)
}
\arguments{
\item{...}{one or more objects which can be interpreted as factors
(including character strings), or a list (or data frame) whose components
 can be so interpreted.}
}
\description{
sets UseNA to 'ifany'.
}

