% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tbl.logic}
\alias{tbl.logic}
\title{create a table, and ensure that the levels TRUE, FALSE and NA are always all
 represented}
\usage{
tbl.logic(x)
}
\arguments{
\item{x}{a logical vector}
}
\description{
create a table, and ensure that the levels TRUE, FALSE and NA are always all
 represented
}

