% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{VtoM}
\alias{VtoM}
\title{transform vector to matrix}
\usage{
VtoM(V, nr = NULL, nc = 2, Ng_odd = FALSE)
}
\arguments{
\item{V}{a vector}

\item{nr}{number of non-empty rows in matrix}

\item{nc}{number of columns in matrix}

\item{Ng_odd}{boolean, apply correction when number of genotyped individuals
is odd}
}
\value{
A n x ncol matrix
}
\description{
transform vector to matrix
}

