% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckGeno.R
\name{CheckGeno}
\alias{CheckGeno}
\title{check GenoM}
\usage{
CheckGeno(GenoM, quiet = FALSE)
}
\arguments{
\item{GenoM}{the genotype matrix}

\item{quiet}{suppress messages}
}
\value{
a list with low call rate individuals (<0.5, 'ExcludedInd') and low
  call rate SNPs (<0.1, 'ExcludedSnps'), if any, which will be excluded from
  pedigree reconstruction.
}
\description{
Check that the provided genotype matrix is in the correct format
}
