% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{CheckLH}
\alias{CheckLH}
\title{check LifeHistData}
\usage{
CheckLH(LifeHistData)
}
\arguments{
\item{LifeHistData}{the dataframe with ID - Sex - Birth year}
}
\value{
a dataframe with corrected formatting of LifeHistData, if required
}
\description{
Check that the provided LifeHistData is in the correct format
}
\keyword{internal}
