% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCompare.R
\name{DyadCompare}
\alias{DyadCompare}
\title{Compare dyads}
\usage{
DyadCompare(PedIN, PedOUT)
}
\arguments{
\item{PedIN}{True pedigree}

\item{PedOUT}{Inferred pedigree}
}
\value{
A 3x3 table with the number of pairs assigned as full siblings (FS),
  half siblings (HS) or unrelated (U) in the two pedigrees.
}
\description{
Count the number of half and full sibling pairs correctly and incorrectly
assigned
}

