% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{LH_HSg5}
\alias{LH_HSg5}
\title{Example life history file}
\format{A data frame with 1000 rows and 3 variables: ID, Sex (1=female,
2=male), and BY (birth year, here cohort)}
\usage{
data(LH_HSg5)
}
\description{
This is the lifehistory file associated with Pedigree II in the paper.
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\references{
Huisman, J. Pedigree reconstruction from SNP data: Parentage
  assignment, sibship clustering, and beyond. (under review) Molecular
  Ecology Resources
}
\seealso{
\code{\link{Ped_HSg5} \link{sequoia}}
}
\keyword{datasets,}
\keyword{sequoia}

