% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkAgePrior.R
\name{MakeAgeprior}
\alias{MakeAgeprior}
\title{Age priors}
\usage{
MakeAgeprior(UseParents = FALSE)
}
\arguments{
\item{UseParents}{Use the age distribution of assigned parents. Otherwise,
equal probabilities across all age differences are assumed.}
}
\value{
A text file with the probability ratio of two individuals being a
  certain type of relative versus being a random draw, conditional on the
  absolute age difference between those individuals.

  One row per absolute age difference (0 - ...), and one column for each
   relationship type, with abbreviations:
   \describe{
     \item{MS}{Maternal siblings}
     \item{PS}{Paternal siblings}
     \item{MGM}{Maternal grandmother}
     \item{PGF}{Paternal grandfather}
     \item{MGF}{Maternal grandfathers and paternal grandmothers}
     \item{UA}{Avuncular}
   }
}
\description{
Calculate age-based prior probabilities for various categories of pairwise
relatives.
}
\details{
if UseParents = TRUE, Retrieve age distributions of maternal & paternal
parents, siblings and grandparents from assigned parents, to use as input
for sibship clustering and grandparent assignment.
If the lifehistory file indicates a single age class, \eqn{MS = PS = 1} and
 \eqn{MGM = PGF = MGF = UA = 0}.
}

