% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedToNum.R
\name{NumToID}
\alias{NumToID}
\title{Change Numeric Pedigree back to Character Pedigree}
\usage{
NumToID(x, k = 0, gID = NULL, DumPrefix = c("F", "M"))
}
\arguments{
\item{x}{vector with numbers.}

\item{k}{1=dam, 2=sire, needed to distinguish dummy females from dummy males.}

\item{gID}{vector with IDs of SNP-genotyped individuals; rownames of
genotype matrix in the exact order.}

\item{DumPrefix}{length-2 character vector to make dummy IDs; length-3 in
case of hermaphrodites.}
}
\value{
A character vector with IDs.
}
\description{
Reverse \code{\link{PedToNum}}, 1 column at a time.
}
\keyword{internal}
