% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckLifeHist.R
\name{orderLH}
\alias{orderLH}
\title{Order Lifehistory Data}
\usage{
orderLH(LH = NULL, gID = NULL)
}
\arguments{
\item{LH}{dataframe with lifehistory information:
\describe{
 \item{ID}{max. 30 characters long,}
 \item{Sex}{1 = females, 2 = males, other numbers = unknown,}
 \item{Birth Year}{(or hatching year) Use negative numbers to denote
 missing values.}
 \item{BY.min}{minimum birth year (optional)}
 \item{BY.max}{maximum birth year (optional)}}}

\item{gID}{character vector with IDs in genotype data, in order of
occurrence.}
}
\value{
A dataframe with the same 5 columns, but with individuals in exactly
  the same order as gID, including padding with 'empty' rows if an individual
  in gID was not in the input-LH. Missing values are recoded to 3 for the
  'Sex' column, and -999 for the birth year columns.
}
\description{
Order lifehistory data to match order of IDs in genotype data,
  filling in gaps with missing values.
}
\keyword{internal}
