% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCompare.R
\name{PedCompare}
\alias{PedCompare}
\title{Compare two Pedigrees}
\usage{
PedCompare(Ped1 = NULL, PedFile1 = NULL, Ped2 = NULL,
  PedFile2 = "PedSeq.txt", sep1 = "\\t", sep2 = "", DumPrefix = c("F0",
  "M0"), FlipCols = FALSE, MergedPed = FALSE)
}
\arguments{
\item{Ped1}{Original pedigree, dataframe with 3 columns.}

\item{PedFile1}{filename of original pedigree, only the first 3 columns
will be used. Provide EITHER Ped1 OR PedFile1.}

\item{Ped2}{Infered pedigree, e.g. SeqOUT$PedSeq}

\item{PedFile2}{filename of infered pedigree, default `PedSeq.txt'}

\item{sep1}{the field separator character in 'PedFile1'}

\item{sep2}{the field separator character in 'PedFile2'}

\item{DumPrefix}{The dummy prefix used. If NULL, the intersect between
the IDs in Pedigrees 1 and 2 is taken as the list of genotyped
individuals. Otherwise, all individuals in Pedigree 2 with IDs
not starting with the Dummy prefix are taken.}

\item{FlipCols}{Are dam & sire columns in the same order (FlipCols=FALSE)
or in flipped order (FlipCols=TRUE) in the two pedigree files?}

\item{MergedPed}{Return combined pedigree? This including three columns
(id.real, dam.real, sire.real) where dummy IDs in Pedigree 2 are replaced
by the most likely non-genotyped individual ID from Pedigree 1.}
}
\value{
when MergedPed=FALSE (default) a 7 x 5 x 2 named numeric array, and when
  MergedPed=TRUE a list with this array and the combined pedigree.

The first dimension of the array denotes the following categories:
\describe{
  \item{RR}{Real genotyped individual, real genotyped parent}
  \item{RD}{Real individual, dummy parent (at least 1 genotyped sibling}
  \item{RT}{Real individual, total}
  \item{DR}{Dummy individual (at least 2 genotyped offspring, real parent
     (i.e., grandparent of the sibship}
  \item{DD}{Dummy individual, dummy parent}
  \item{DT}{Dummy total}
  \item{TT}{Total total, includes all genotyped & non-genotyped individuals
     in the pedigree}
}
The second dimension gives the outcomes:
\describe{
  \item{Total}{The total number of individuals with a parent assigned in
    either or both pedigrees}
  \item{Match}{The same parent is assigned in both pedigrees (non-missing).
     For dummy parents, it is considered a match if the inferred sibship
     containing the most offspring of a non-genotyped parent, consists for
     more than half of its offspring.}
  \item{Mismatch}{Different parents assigned in the two pedigrees. When
   a sibship according to Pedigree 1 is split over two sibships in Pedigree
   2, the smaller fraction is included in the count here.}
  \item{P1only}{Parent in Pedigree 1 but not 2; includes non-assignable
    parents.}
  \item{P2only}{Parent in Pedigree 2 but not 1.}
}
The third dimension separates between maternal and paternal assignments,
where e.g. paternal 'DR' is the assignment of fathers to both maternal and
paternal sibships.
}
\description{
Compare an inferred pedigree to a previous or simulated pedigree,
including comparison of sibship clusters and sibship grandparents.
}
\details{
Provide either a dataframe of the pedigree, or the filename of a text file.
In both cases, the inferred pedigree 2 is assumed to have the columns id-
dam-sire, and Pedigree 1 may be in the same order (FlipCols=FALSE, default)
or have order id-sire-dam, in which case FlipCols=TRUE should be used.
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\seealso{
\code{\link{sequoia}}
}

