% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkAgePrior.R
\name{PlotAgePrior}
\alias{PlotAgePrior}
\title{Plot age priors}
\usage{
PlotAgePrior(AP = NULL, GPlines = TRUE)
}
\arguments{
\item{AP}{matrix with age priors of dimension nAgeclasses rows by 9 columns,
OR a list returned by \code{\link{MakeAgePrior}} when called with
Return='all', OR a list returned by \code{\link{sequoia}} containing an
element 'AgePriors'.}

\item{GPlines}{show the lines for grandparental & avuncular relationships?
Can sometimes make the plot overly crowded.}
}
\description{
visualise the age-difference based prior probability ratios
}
\details{
When called with an output list from \code{MakeAgePrior}, the raw
  distributions are shown too in a separate panel, in addition to the the
  scaled, flattened & smoothened distributions used for pedigree
  reconstruction.
}
\seealso{
\code{\link{MakeAgePrior}}, \code{\link{SummarySeq}}
}
