% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkAgePrior.R
\name{PlotAgePrior}
\alias{PlotAgePrior}
\title{Plot age priors}
\usage{
PlotAgePrior(AP = NULL, legend = TRUE)
}
\arguments{
\item{AP}{matrix with age priors (P(A|R)/P(A)) with age differences in rows
and relationships in columns; by default M: maternal parent (mother), P: paternal
parent (father), FS: full siblings, MS: maternal siblings (full + half),
PS: paternal siblings.}

\item{legend}{if TRUE, a new plotting window is started and
\code{\link{layout}} is used to plot a legend next to the main plot. Set to
FALSE if you want to add it as panel to an existing plot (e.g. with
par(mfcol=c(2,2))).}
}
\value{
a heatmap
}
\description{
visualise the age-difference based prior probability ratios as a
  heatmap
}
\examples{
\dontrun{
PlotAgePrior(SeqOUT$AgePriors)
}

}
\seealso{
\code{\link{MakeAgePrior}}, \code{\link{SummarySeq}}
}
