% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_F90wrappers.R
\name{SeqDup}
\alias{SeqDup}
\title{Check data for duplicates.}
\usage{
SeqDup()
}
\value{
A dataframe with the numbers of duplicated genotypes, duplicated IDs in the genotype file, number of duplicated IDs in the life history file, and number of individuals with unknown sex.
}
\description{
Check the genotype and life history files for duplicate IDs (not permitted)
and duplicated genotypes (not advised), and count how many individuals in
the genotype file are included in the life history file (permitted).
}
\details{
The filenames are taken from the SequioiaSpecs text file, generated by
\code{\link{SeqPrep}}. The order of IDs in the genotype and life history file
is not required to be identical, and not all individuals in the genotype file
are required to be in the life history file.
}

