% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_F90wrappers.R
\name{SeqPrep}
\alias{SeqPrep}
\title{Convert genotype file if needed, and save parameter values.}
\usage{
SeqPrep(GenoFile = NULL, GenoFormat = "seq", LifeHistFile = NULL,
  Err = 1e-04, MaxMismatch = 3, Tfilter = -2, Tassign = 0.5,
  MaxSibshipSize = 100, NumSibRounds = 3, Complexity = "full",
  AgePriorFile = "AgePriors.txt", ParentageFile = "ParentsAssigned.txt",
  PedigreeFile = "PedSeq.txt")
}
\arguments{
\item{GenoFile}{character string with name of genotype file}

\item{GenoFormat}{One of "raw", "ped", "col" or "seq", see ?GenoConvert.}

\item{LifeHistFile}{Life history data, with 3 columns:
ID: max. 30 characters long,
Sex: 1 = females, 2 = males, other numbers = unkown,
BirthYear: Use negative numbers to denote missing values. If the species
has multiple generations per year, use an integer coding such that the
candidate parents' `Birth year' is at least one larger than their putative
offspring.}

\item{Err}{Estimated genotyping error rate.}

\item{MaxMismatch}{Maximum number of loci at which candidate parent and
offspring are allowed to be opposite homozygotes, or be excluded.}

\item{Tfilter}{Threshold log-likelihood ratio between a proposed
relationship versus unrelated, to select candidate relatives. Typically a
negative value, related to the fact that unconditional likelihoods are
calculated during the filtering steps. More negative values may decrease
non-assignment, but will increase computational time.}

\item{Tassign}{Minimum log-likelihood ratio required for acceptance of
proposed relationship, relative to next most likely relationship. Higher
values result in more conservative assignments.}

\item{MaxSibshipSize}{Maximum number of offspring for a single individual
(a generous safety margin is advised).}

\item{NumSibRounds}{Number of iterations of sibship clustering.}

\item{Complexity}{Either "full" (default), "simp" (no explicit consideration
of inbred relationships) or "mono" (monogamous breeding system)}

\item{AgePriorFile}{Filename for ageprior (in & out, if applicable)}

\item{ParentageFile}{Filename for assigned parents}

\item{PedigreeFile}{Filename for pedigree after sibship clustering}
}
\value{
A dataframe with the number of individuals found in the genotype file and in the lifehistory file, and the number of SNPs scored.
}
\description{
Write the user-specified parameter values to the (temporary) file
`SequoiaSpecs.txt'. If GenoFormat="raw" or "col", call \code{\link{GenoConvert}}.
}
\details{
Please do not add or delete rows in the SequoiaSpecs text file. Changing
 values is possible, but increasing the number of SNPs or individuals beyond
 the numbers present in the files, or refering to non-existing files, will
 cause errors and may cause R to crash.
}

