% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequoia_Main.R
\name{sequoia}
\alias{sequoia}
\title{Pedigree Reconstruction}
\usage{
sequoia(GenoFile = NULL, GenoFormat = "seq", LifeHistFile = NULL,
  Prep = TRUE, CheckDup = TRUE, Parentage = TRUE, AgePriors = "par",
  Sibships = TRUE, SeqList = NULL, Err = 1e-04, MaxMismatch = 3,
  Tfilter = -2, Tassign = 0.5, MaxSibshipSize = 100, NumSibRounds = 5,
  Complexity = "full", WriteFiles = FALSE, Dir = NULL,
  AgePriorFile = "AgePriors.txt", ParentageFile = "ParentsAssigned.txt",
  PedigreeFile = "PedSeq.txt", quiet = FALSE)
}
\arguments{
\item{GenoFile}{character string with name of genotype file. Assumed to be in
current working directory unless complete path is provided.}

\item{GenoFormat}{One of "raw", "ped", "col" or "seq". If "seq", genotype
file should have an ID column and 1 column per SNP, coded as 0, 1, 2 or -9
 (missing), and no header row. For other formats, see \code{\link{GenoConvert}}.}

\item{LifeHistFile}{character string with name of life history data file,
with 3 columns:
\itemize{
\item ID: max. 30 characters long,
\item Sex: 1 = females, 2 = males, other numbers = unkown,
\item BirthYear: Use negative numbers to denote missing values.}
If the species has multiple generations per year, use an integer coding
 such that the candidate parents' `Birth year' is at least one larger than
 their putative offspring.}

\item{Prep}{TRUE: Save current parameter values to file, FALSE: use
previous settings, as stored in `SequoiaSpecs.txt'.}

\item{CheckDup}{Check the genotype and life history files for duplicate IDs
and duplicated genotypes (neither is advised, but will not generate an
error), and count how many individuals in the genotype file are not
included in the life history file (permitted, but may hinder pedigree
 inference).}

\item{Parentage}{Run parentage analysis; assign genotyped parents to
genotyped individuals.}

\item{AgePriors}{Calculate age-based prior probabilities for various
categories of pairwise relatives, based on the age distribution of
assigned parents (``par'', default once parents have been assigned),
or use a default non-informative prior (``flat''), or use the ageprior
 from a previous run (``old''), stored as `AgePriors.txt' or provided in
  SeqList).}

\item{Sibships}{Perform sibship clustering, including assignment of
grandparents to sibships and avuncular relationships between sibships.
Note that this is by far the most time consuming step and may take several
hours.}

\item{SeqList}{list with output from a previous partial run, see example.}

\item{Err}{Estimated genotyping error rate.}

\item{MaxMismatch}{Maximum number of loci at which candidate parent and
offspring are allowed to be opposite homozygotes (integer).}

\item{Tfilter}{Threshold log-likelihood ratio between a proposed
relationship versus unrelated, to select candidate relatives. Typically a
negative value, related to the fact that unconditional likelihoods are
calculated during the filtering steps. More negative values may decrease
non-assignment, but will increase computational time.}

\item{Tassign}{Minimum log-likelihood ratio required for acceptance of
proposed relationship, relative to next most likely relationship. Higher
values result in more conservative assignments.}

\item{MaxSibshipSize}{Maximum number of offspring for a single individual
(a generous safety margin is advised).}

\item{NumSibRounds}{Number of iterations of sibship clustering. Sibship
clustering is iterated until convergence or until this maximum is reached.}

\item{Complexity}{Either "full" (default), "simp" (simplified, no explicit
consideration of inbred relationships) or "mono" (monogamous).}

\item{WriteFiles}{if TRUE, save all (intermediate) output to text files,
if FALSE (default), return results as a list}

\item{Dir}{Directory to which output files will be written. If NULL
(default) and WriteFiles=TRUE a subdirectory 'sequoiaOUT' in the current
 working directory is created, if NA use current working directory.}

\item{AgePriorFile}{Filename for ageprior (in & out, if applicable)}

\item{ParentageFile}{Filename for assigned parents}

\item{PedigreeFile}{Filename for pedigree after sibship clustering}

\item{quiet}{suppress messages}
}
\value{
A list with the following components when WriteFiles=FALSE, or
 otherwise a set of textfiles with the following names:
\tabular{lll}{
\strong{Name} \tab \strong{Generated by} \tab \strong{Description} \cr
AgePriors \tab Prep, \tab Age-based prior probabilities \cr
\tab AgePriors == "par" \tab \cr
DummyParents \tab Sibships \tab Half-sib clusters \cr
DuplicatesFound \tab CheckDup \tab duplicated genotypes and IDs \cr
NumberParents \tab Parentage \tab No. assigned parents \cr
ParentsAssigned \tab Parentage \tab Provisionary pedigree \cr
ParentsTotLik \tab Parentage \tab Total likelihood per iteration \cr
PedSeq \tab Sibships \tab Pedigree \cr
SequoiaSpecs \tab Prep \tab Parameter values \cr
SibTotLik \tab Sibships \tab Total likelihood per iteration \cr
UnassignedParents \tab Parentage \tab Non-assigned likely PO pairs \cr
UnassignedSibs \tab Sibships \tab Non-assigned likely relatives
}

ParentsAssigned and PedSeq have the following columns:
  \item{id}{Individual ID}
  \item{dam}{Assigned mother, or NA}
  \item{sire}{Assigned father, or NA}
  \item{LLR_dam}{Log10-Likelihood Ratio (LLR) of this female being the mother,
     versus the next most likely relationship between the focal individual and
     this female (see Details for relationships considered)}
  \item{LLR_sire}{idem, for male parent}
  \item{LLR_pair}{LLR for the parental pair, versus the next most likely
    configuration (containing one or neither parent)}
  \item{OH_dam}{Number of loci at which the offspring and mother are
     opposite homozygotes}
  \item{OH_sire}{idem, for male parent}

For details on the other elements, see the vignette.
}
\description{
Perform pedigree reconstruction based on SNP data, including parentage
assignment and sibship clustering.
}
\details{
For each pair of candidate relatives, the likelihoods are calculated of them
 being parent-offspring, full siblings, half siblings, grandparent-
 grandoffspring, full avuncular (niece/nephew - aunt/uncle), half avuncular/
 great-grandparental/cousins, or unrelated. Assignments are made if the
 likelihood ratio (LLR) between the focal relationship and the most likely
 alternative exceed the threshold Tassign

Detailed explanation of the various options and interpretation of the output
 is provided in the vignette.
}
\examples{
# Reconstruct a 2-generation pedigree;s See vignette for details.

curdir <- getwd()
setwd(tempdir())
# copy files included in package (note: does not work without copying)
file.copy(system.file(file.path("extdata",c("SimGeno_example.txt",
   "LH_HSg5.txt")), package="sequoia"), tempdir())

# save parameter values & do parentage assignment
SeqOUT <- sequoia(GenoFile = "SimGeno_example.txt",
                  LifeHistFile = "LH_HSg5.txt",
                  Prep=TRUE, Parentage=TRUE, Sibships = FALSE)
names(SeqOUT)
SeqOUT$NumberParents

# do sibship clustering; slow (minutes to hours) for large datasets,
# but fast (seconds) for this example dataset with 214 individuals.
SeqOUT <- sequoia(Prep=FALSE, SeqList=SeqOUT, Parentage=FALSE,
                  Sibships=TRUE)
names(SeqOUT)
head(SeqOUT$PedSeq)  # pedigree with parents + dummy parents of sibships
file.remove("SimGeno_example.txt", "LH_HSg5.txt")
setwd(curdir)

# Check for mismatches with input pedigree:
compare <- PedCompare(PedFile1=system.file(file.path("extdata","Ped_HSg5.txt"),
      package="sequoia"), Ped2=SeqOUT$PedSeq, MergedPed=TRUE)
compare$Score   # 2 mismatches, due to genotyping errors

}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\references{
Huisman, J. Pedigree reconstruction from SNP data: Parentage
  assignment, sibship clustering, and beyond. (under review) Molecular
  Ecology Resources
}
\seealso{
\code{\link{GenoConvert}, \link{PedCompare}}
}

