% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCompare.R
\name{SimGeno}
\alias{SimGeno}
\title{Simulated genotypes}
\usage{
SimGeno(PedFile = "TruePed.txt", MafFile = NA, OutFile = "SimGeno.txt",
  nSnp = 400, nGen = 10, PropLQ = 0, MisHQ = 0.005, MisLQ = 0.3,
  ParMis = 0.2, ErHQ = 0.005, ErLQ = 0.005, quiet = FALSE)
}
\arguments{
\item{PedFile}{Name of the pedigree file, with columns ID - dam - sire}

\item{MafFile}{(optional) file with minor allele frequency at each locus.
If none specified, allele frequencies will be sampled from a uniform
distribution between 0.3 and 0.5}

\item{OutFile}{filename for simulated genotypes}

\item{nSnp}{number of SNPs to simulate}

\item{nGen}{maximum number of generations to consider}

\item{PropLQ}{proportion of low-quality samples}

\item{MisHQ}{average missingness for high-quality samples, assuming a
beta-disstribution with alpha = 1.}

\item{MisLQ}{average missingness in low-quality samples}

\item{ParMis}{Proportion of parents with fully missing genotype}

\item{ErHQ}{error rate in high quality samples. Default value is
0.005, following Anderson & Garcia 2006.}

\item{ErLQ}{error rate in low quality samples}

\item{quiet}{suppress messages}
}
\value{
a file with genotype data encoded as 0/1/2/-9, in a format
  similar to that generated by Plink's --recodeA option.
}
\description{
Simulate SNP genotype data from a pedigree, with optional missingess and
 errors.
}
\details{
Please ensure the pedigree is a valid pedigree, for example by first running
fixPedigree() from library Pedantics.
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}

