% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{SimGeno_example}
\alias{SimGeno_example}
\title{Example genotype file}
\format{
A genotype matrix with 214 rows (ids) and 200 columns (SNPs). Each
  SNP is coded as 0/1/2 copies of the reference allele, with -9 for missing
  values. Ids are stored as rownames.
}
\usage{
data(SimGeno_example)
}
\description{
Simulated genotype data for cohorts 1+2 in Pedigree Ped_HSg5
}
\seealso{
\code{\link{Ped_HSg5}, \link{SimGeno}}
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
\keyword{datasets}
\keyword{sequoia}
