% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteToFiles.R
\name{writeColumns}
\alias{writeColumns}
\title{Write Data to a File Column-wise}
\usage{
writeColumns(x, file = "", row.names = TRUE, col.names = TRUE)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame.
If not, it is attempted to coerce x to a matrix.}

\item{file}{a character string naming a file.}

\item{row.names}{a logical value indicating whether the row names of x are
to be written along with x.}

\item{col.names}{a logical value indicating whether the column names of x
are to be written along with x.}
}
\description{
Write data.frame or matrix to a text file, using white
space padding to keep columns aligned as in \code{print}.
}
